add websocket message event handling
