+++ server/test-client/index.js
@@ -1,33 +1,37 @@
 const WebSocket = require('ws');
-const readline = require('readline');
 const logger = require('../src/logger');
+const { setupRL, getRL } = require('./cli');
 
 const WS_PATH = 'ws://localhost:8080';
 const ws = new WebSocket(WS_PATH);
-const rl = readline.createInterface({
-  input: process.stdin,
-  output: process.stdout,
-});
-
-rl.setPrompt('send> ');
-rl.on('line', (line) => {
-  if (line === 'exit') {
-    rl.close();
-  } else {
-    ws.send(line);
-    rl.prompt();
-  }
-}).on('close', () => {
-  ws.close();
-  logger.info('bye');
-  process.exit(0);
-});
 
 ws.on('open', () => {
   logger.info(`Connection opened with at ${WS_PATH}`);
-  rl.prompt();
+  getRL().prompt();
 });
 
 ws.on('message', (data) => {
   logger.info(`Incoming data:\n${data}`);
+  const parsed = JSON.parse(data);
+  switch (parsed.type) {
+    case 'plr.move':
+      logger.info(`player ${parsed.plr.id} made move ${parsed.move}`);
+      break;
+    case 'plr.join':
+      logger.info(`player ${parsed.plr.id} joined`);
+      break;
+    case 'plr.disconnect':
+      logger.info(`player ${parsed.plr.id} disconnected`);
+      break;
+    case 'game.end':
+      logger.info(`game ended, ${parsed.winner.id} won!`);
+      break;
+    case 'game.start':
+      logger.info(`game start, initial state ${parsed.initialState}`);
+      break;
+    default:
+      logger.warn(`No type for\n${parsed}`);
+  }
 });
+
+setupRL(ws);
